requirements.txt
# Dependências principais
google-generativeai>=0.3.0
pymongo>=4.0.0
python-dotenv>=1.0.0

# Processamento de dados
pandas>=1.5.0
numpy>=1.21.0

# Utilitários
pathlib2>=2.3.0
requests>=2.28.0

# Automação web (para revisão com Gemini)
selenium>=4.0.0
pyperclip>=1.8.0

# Logging e configuração
logging>=0.4.9.6

# Para desenvolvimento (opcional)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Para testes (opcional)
pytest>=7.0.0
pytest-cov>=4.0.0 